backend/requirements.txt
fastapi
uvicorn[standard]
ctransformers
chromadb
sentence-transformers
pypdf
numpy
python-multipart
beautifulsoup4
python-docx